Add histograms object and constructors
The structure type is an AOS of value-pairs of the form (counts,
logdos). Here the AOS makes sense since the histogram value and
the logdos value are used in the same step of each iteration. This
should lead to a reduction in the cache misses in the REWL
simulaiton. Note the SOA will separate the histogram count and the
the logdos value by the bin size.