Start Ising2d Hamiltonian
Setting up some Hamiltonian functionality. All parameter text files
will be .cxx files and contain useful constexpr (such as the size,
etc.). This .cxx file will be modified by the user before compile
time. It is then included as a header.

The observables are templated and have an enum class. These obser-
vables are "non-energetic," meaning that they are not determined
by the Wang Landau histogram process. Instead, they are computed
as the mean value of each observable within each bin. This is a
pretty big choice and may be modified later on. ONE OBSERVABLE
MUST ALWAYS BE THE COUNTS_PER_BIN. This is the number of times
the simulation has entered a specific bin throughout the entire
runtime. This is used for quick updates to the observables and
makes sure the REWL arrays need not mix with the Hamiltonian
arrays for memory reasons and Hamiltonian generalization reasons.

Moving forward, the actual Hamiltonian needs to be templated and
defined for inclusion and testing in the REWL functions.