Rearrange header layouts
Added compiler flags too, as well as an option to specify the model
Hamiltonian -- right now the default is the 2d Ising model.

The way this is going to work for different models is to then add
compiler directives in the model_hamiltonians.hpp file and add
options to the parent CMakeLists.txt to include different models.
Then the template type Hamiltonian will be called in the Wang Landau
functions after being set inside of model_hamiltonians.hpp. For
example, if -DISING2D=ON then using Hamiltonian = Ising2d<data_t>
is set.

The pileup of headers due to constexpr parameters are starting to
get a little dubious and hard to track. It may be helpful to
cascade some subdirectories to help with this, but then this hides
the .cxx parameters file under a mountain of directories so I'm
not sure at the moment.