Add util a folder for special math functions
Right now it houses a wrapper around standard library random
number generators. This is to separate the random number generation
from the Wang Landau functionality. However, the each walker will
need its own random number generator because they are not thread
safe. This may lead to a performance hit for single-node simulations
with shared memory spaces.

Also the util folder is added to the include path.