Found a couple of bugs
This included the tolerance being messed up before. Now it works
and the simulation runs incredibly quickly! Something else is wrong
though dealing with the flatness not being changed after the first
step. So if it's flat once, it's flat the whole time after (even
when it's not actually flat).