Concatenation works with equally-sized windows
The new function is simpler and iterative. It concatenates all
windows beyond the first with the final vectors, and so it will
allow for multiple overlapping windows per bin without loss of
data. One still needs to be cautious when using this to be sure
the bins are representable as a float and there are an integer
amount of bins per window. This ultimately put a constraint on
the number of windows and the size of the system, and so must be
active in guaranteeing the chosen overlap works.