WIP: Need to implement the glazier after the fact
The Hamiltonian with the knowledge of the ground state needs to
distribute its ground state energy to everyone so that everyone's
glazier function can then set up the histograms and binnings.
After this, then the glazier needs to be used to reinitialize the
histograms and histogram functor, etc.

This is kind of convoluted unfortunately since the Hamiltonian and
and the glazier are separate. Could it be solved if I move the
glazier inside the Hamiltonian? I'm sure this would lead to other
problems though...