Not what I thought previously
Randomizing the sites didn't do anything but slow down the annealing
process by a factor of 2 - 3. So going back to the fast way.

I had to go to monstrously large fields for 16 x 16 to see breakup.
h >= 10 for this L. For 32 x 32, h >= 5.

Increasing the number of temperatures seemed to help.