Still testing but the annealing doesn't need to take that long
A factor of 100x speedup was achieved over before simply by reducing
the total number of sweeps between temperatures down from 10^5 to
10^3. It seems like this is a happy medium. This also allows for
more temperatures to be measured, and so the simulation is more
adiabatic.

With this I saw breakup with L = 16, 32 at h = 5 and for L = 48 at
h = 3. Didn't see breakup for L = 48, 64 at h = 2.