Error with multiple replicas and observables fixed
Not sure why yet, but if the observables are not
averaged after each step and are instead only
averaged at the very end of the simulation in each
window, then there doesn't appear to be any problems
with them. Somehow the logic is wrong for these,
but not for the logDoS? Also somehow the logic is
worse for quickly changing observables and not
slowly changing ones? I don't have answers to these
questions right now.