Change rng to double
This is a hunch since on the last step with RFIM, it takes so much
longer to finish when the logdos increment is 1.2 x 10^-7. I think
the choice of float was basically underflowing the registers so
every move was guaranteed and this was making it very hard for the
energy histograms to be flat.