WIP: Add Ashkin-Teller density plots
These density plots will be 2d histograms in the sigma-tau plane
that will keep track of the distribution of sigmas and taus for
every energy bin as the simulation proceeds.

Only the microcanonical plots will be written to disk. Additional
processing software will be necessary to compute the average
density at a given temperature.