Starting RFIM with dipolar interactions
These are specialized dipolar interactions in that they have
and angular component as well coming from their phononic-
nematic source. Hopefully the details will be documented
somewhere along the way.

Here I've written vector operations on a 2d periodic grid
necessary for defining the long-range interactions in the
Hamiltonian as

    \sum_{ij} C_{ij} \sigma_i \sigma_j