Modify simulated annealing routine to be adaptive
Now the simulated annealing routine will only end an iteration when
the system is equilibrated. This is defined by an adaptive pairing
of block averages of the energies, and the system is considered
equilibrated if either of the following statements are true:

    * The energy variance within the latter block is zero.

    * The average energy of the second block falls within a
      standard deviation of the average in the first block, AND
      the energy variances from each block are within some
      tolerance of once another.

The first condition is simply that the system is frozen. If frozen,
there is no need to continue the annealing process at this partic-
ular temperature.

The second contition is stating that the blocking average of
energies is time-independent. This checks whether the average
energy within the block is time-independent and whether the size
of the specific heat (energy variance) agrees between blocks.