Address Book data structures are being implemented
This houses all functionality relating the different lattice sites
to one another. So far, there is a Lattice_Address_Book abstract
class and a derived class for a 2D square periodic lattice.

The src/Hamiltonians/grid folder as of now is deprecated. I will
remove it soon.

There is also the square_2d_grid_helpers.hpp file that defines
helpful function in the lattice, with identical compile-time and
run-time variants based on preprocessor macros. These are very
generalizable for hypercubic lattices. Not so sure how it will
work for stranger lattices.