Performance takes a hit with the new neighbor method
This is to be expected considering that I added a beefier data
structure. It's very likely that if I change this data structure
to simply an object that calls functions from a function array,
then I'll get a performance boost due to fewer cache misses.

Overall, not too shabby for a first pass.